requirements.txt
streamlit
google-generativeai
Pillow
requests
beautifulsoup4
pypdf